HKILA bot/requirements.txt
streamlit
openai
google-auth
google-api-python-client
PyPDF2
langchain
faiss-cpu